+++ src/test/java/de/geofabrik/railway_routing/parsers/RailAverageSpeedParserTest.java
@@ -7,7 +7,6 @@ import com.graphhopper.reader.ReaderWay;
 import com.graphhopper.routing.ev.ArrayEdgeIntAccess;
 import com.graphhopper.routing.ev.DecimalEncodedValue;
 import com.graphhopper.routing.ev.EdgeIntAccess;
-import com.graphhopper.routing.ev.EncodedValueLookup;
 import com.graphhopper.routing.ev.FerrySpeed;
 import com.graphhopper.routing.ev.MaxSpeed;
 import com.graphhopper.routing.ev.TurnCost;
@@ -15,22 +14,16 @@ import com.graphhopper.routing.ev.VehicleAccess;
 import com.graphhopper.routing.ev.VehicleSpeed;
 import com.graphhopper.routing.util.EncodingManager;
 import com.graphhopper.storage.IntsRef;
-import com.graphhopper.util.PMap;
-
-public class RailSpeedParserTest {
+public class RailAverageSpeedParserTest {
     
     DecimalEncodedValue speedEnc;
     RailAverageSpeedParser parser;
     
-    public RailSpeedParserTest() {
-        parser = createSpeedParser(createEncodingManager(), new PMap());
+    public RailAverageSpeedParserTest() {
+        parser = new RailAverageSpeedParser(createEncodingManager());
     }
 
-    private static RailAverageSpeedParser createSpeedParser(EncodedValueLookup lookup, PMap properties) {
-        return new RailAverageSpeedParser(lookup, properties);
-    }
-
-    private EncodingManager createEncodingManager() {
+    protected EncodingManager createEncodingManager() {
         speedEnc = VehicleSpeed.create("rail", 6, 5, true);
         return new EncodingManager.Builder()
                 .add(VehicleAccess.create("rail"))